Initial commit of v1.0 of AMI to HPCTrainingDock AMI is usable to generate functional machine images for use with local raw image installs, USB disks, ISOs, and kvm/pxeboot type images with the tooling one needs to be functional with the tooling in the HPCTrainingDock repository.
This is an initial commit, and we will be leveraging the installation
scripts for various tools as indicated in this repository as well
during full integration.  As if now, we can generate non-integrated
bootable raw, qcow2, and kvm (kernel and an initramfs-ramboot) images.

Subsequent commits will provide for the integration.